+++ exoscale/datasource_exoscale_elastic_ip.go
@@ -36,7 +36,7 @@ const (
 
 func dataSourceElasticIP() *schema.Resource {
 	return &schema.Resource{
-		Description: `Fetch Exoscale [Elastic IPs (EIP)](https://community.exoscale.com/documentation/compute/eip/) data.
+		Description: `Fetch Exoscale [Elastic IPs (EIP)](https://community.exoscale.com/product/networking/eip/) data.
 
 Corresponding resource: [exoscale_elastic_ip](../resources/elastic_ip.md).`,
 		Schema: map[string]*schema.Schema{