+++ pkg/resources/block_storage/resource_volume.go
@@ -26,7 +26,7 @@ import (
 	"github.com/exoscale/terraform-provider-exoscale/pkg/utils"
 )
 
-const ResourceVolumeDescription = `Manage [Exoscale Block Storage](https://community.exoscale.com/documentation/block-storage/) Volume.
+const ResourceVolumeDescription = `Manage [Exoscale Block Storage](https://community.exoscale.com/product/storage/block-storage/) Volume.
 
 Block Storage offers persistent externally attached volumes for your workloads.
 `