+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 IMPROVEMENTS:
 - compute_instance: make disk_size a required field #419
 - docs: note that domain_record.content format depnds on record type #429
+- sks-cluster: support for feature-gates #431
 
 BUG FIXES: